models/retinanet/requirements.txt

# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# model\callbacks.py: 1
# model\common.py: 1
# model\image.py: 5
# model\initializers.py: 3
# model\loss.py: 1
# model\misc.py: 1
# model\optimizer.py: 1
# model\pascal_voc.py: 8
# model\resnet.py: 1
# model\retinanet.py: 1
# show_features.py: 8,9,10
# training.py: 4
Keras == 2.1.6

# model\image.py: 3
# model\pascal_voc.py: 10
Pillow == 5.0.0

# model\resnet.py: 2,3
keras_resnet == 0.1.0

# model\pascal_voc.py: 11
# show_features.py: 1
# test.py: 1,14
matplotlib == 2.1.2

# debug.py: 2
# model\anchors.py: 1
# model\image.py: 6
# model\initializers.py: 4
# model\misc.py: 2
# model\pascal_voc.py: 9
# model\retinanet.py: 2
# model\transform.py: 3
# model\visualization.py: 2
# show_features.py: 2
# test.py: 11
# test_webcam.py: 4
numpy == 1.14.2

# model\pascal_voc.py: 12
six == 1.11.0

# model\tensorflow_backend.py: 1
tensorflow_gpu == 1.7.0

opencv-python == 3.4.0
